Finish styling for HeaderMiddle
Because:
- It needed to be finished.

Implemented the cart items display to match
the cart title width so the display remains
beautiful regardless of the lenght of display.

Also refactored the HeaderStrip IconLinks to use css flex.
Previous implementation did not correctly align the text and icons
vertically.

Explain exactly what was done in this commit with more depth than the
50 character subject line. Remember to wrap at 72 characters!